system_orchestrator/requirements.txt
# System Orchestrator Dependencies
# Windows 11 + Windsurf + Python System-level Orchestration

# Core
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Monitoring
psutil>=5.9.0

# Networking
httpx>=0.25.0
requests>=2.31.0
openai>=1.0.0

# Security
cryptography>=41.0.0
keyring>=24.0.0

# Console/Rich output
rich>=13.7.0

# Windows-specific (conditional)
pywin32>=306; sys_platform == 'win32'
WMI>=1.5.1; sys_platform == 'win32'

# Optional: PyInstaller for deployment
pyinstaller>=6.0.0

# Optional: Additional utilities
asyncio-mqtt>=0.16.0
aiofiles>=23.2.0
